# media-service/swagger.yml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'http://swagger.io/terms/'
  contact: { }
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8090'
basePath: /
tags:
  - name: tar-file-controller
    description: Tar File Controller
paths:
  /api/v1/tarFile:
    post:
      tags:
        - tar-file-controller
      summary: saveTarFile
      operationId: saveTarFileUsingPOST
      consumes:
        - multipart/form-data
      produces:
        - '*/*'
      parameters:
        - name: tarFile
          in: formData
          description: tarFile
          required: true
          type: file
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/api/v1/tarFile/{fileName}':
    get:
      tags:
        - tar-file-controller
      summary: getTarFile
      operationId: getTarFileUsingGET
      produces:
        - '*/*'
      parameters:
        - name: fileName
          in: path
          description: fileName
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /api/v1/tarFiles:
    get:
      tags:
        - tar-file-controller
      summary: getAllTarFiles
      operationId: getAllTarFilesUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    delete:
      tags:
        - tar-file-controller
      summary: deleteAllTarFiles
      operationId: deleteAllTarFilesUsingDELETE
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false